pyproject.toml
[project]
name = "pixelflut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
]
